feat: checking access + some extensions

+++ src/Controllers/AppointmentController.cs
@@ -1,7 +1,10 @@
 using Contracts.Appointment;
 using Contracts.Appointment.AppointmentHelpers;
 using Contracts.Organization;
+using Contracts.User;
 using Domain.Appointment;
+using Exceptions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -9,9 +12,11 @@ namespace Controllers;
 
 [ApiController]
 [Route("api/v1/appointment")]
+[Authorize]
 public class AppointmentController : ControllerBase
 {
     private readonly IAppointmentService _appointmentService;
+    private readonly IPermissionService _permissionService;
 
     public AppointmentController(IAppointmentService appointmentService)
     {
@@ -22,10 +27,15 @@ public class AppointmentController : ControllerBase
     /// Creates an appointment.
     /// </summary>
     [HttpPost]
+    [Authorize(Roles = "Admin,SeniorManager,Manager")]
     [Produces("application/json")]
     [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(AppointmentDetails))]
     public async Task<IActionResult> Create([FromBody]CreateAppointmentRequest request, CancellationToken cancellationToken)
     {
+        if (!(await _permissionService.CheckAdminOrManagerInOrganizationBySlotId(User.Identity.Name, request.SlotId)))
+        {
+            return Forbid();
+        }
         var appointmentDetails = await _appointmentService.CreateAppointmentAsync(request, cancellationToken);
         
         return CreatedAtAction(nameof(GetAppointmentById), new { id = appointmentDetails.Id }, appointmentDetails);
@@ -40,6 +50,10 @@ public class AppointmentController : ControllerBase
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(AppointmentDetails))]
     public async Task<IActionResult> GetAppointmentById([FromRoute] int id, CancellationToken cancellationToken)
     {
+        if (!(await _permissionService.CheckAppointmentPermissionAccess(User.Identity.Name, id)))
+        {
+            return Forbid();
+        }
         var appointmentDetails = await _appointmentService.GetAppointmentByIdAsync(id, cancellationToken);
         
         return Ok(appointmentDetails);
@@ -53,7 +67,19 @@ public class AppointmentController : ControllerBase
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(List<AppointmentDetails>))]
     public async Task<IActionResult> GetAppointments(CancellationToken cancellationToken)
     {
-        var appointments = await _appointmentService.GetAppointmentsAsync(cancellationToken);
+        // admin - all, manager - only in organization, user - only where him
+        int? organizationId = await _permissionService.GetOrganizationId(User.Identity.Name);
+        int? userId = null;
+        if (await _permissionService.HasUserRole(User.Identity.Name))
+        {
+            userId = await _permissionService.GetUserId(User.Identity.Name);
+        }
+        
+        var request = GetAllAppointmentsRequest.Builder
+            .WithOrganizationId(organizationId)
+            .WithUserId(userId)
+            .Build();
+        var appointments = await _appointmentService.GetAppointmentsAsync(request, cancellationToken);
         
         return Ok(appointments);
     }
@@ -63,12 +89,17 @@ public class AppointmentController : ControllerBase
     /// </summary>
     /// <param name="id"></param>
     [HttpPut("user/add/{id}")]
+    [Authorize(Roles = "Admin,SeniorManager,Manager")]
     [Produces("application/json")]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(AppointmentDetails))]
     public async Task<IActionResult> AddUser(int id, int userId, CancellationToken cancellationToken)
     {
-        var appointmentDetails = await _appointmentService.AddUserToAppointmentAsync(id, userId, cancellationToken);
+        if (!(await _permissionService.CheckAppointmentPermissionAccess(User.Identity.Name, id)))
+        {
+            return Forbid();
+        }
         
+        var appointmentDetails = await _appointmentService.AddUserToAppointmentAsync(id, userId, cancellationToken);
         return Ok(appointmentDetails);
     }
     
@@ -77,36 +108,53 @@ public class AppointmentController : ControllerBase
     /// </summary>
     /// <param name="id"></param>
     [HttpPut("user/remove/{id}")]
+    [Authorize(Roles = "Admin,SeniorManager,Manager")]
     [Produces("application/json")]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(AppointmentDetails))]
     public async Task<IActionResult> RemoveUser(int id, int userId, CancellationToken cancellationToken)
     {
-        var appointmentDetails = await _appointmentService.RemoveUserFromAppointmentAsync(id, userId, cancellationToken);
+        if (!(await _permissionService.CheckAppointmentPermissionAccess(User.Identity.Name, id)))
+        {
+            return Forbid();
+        }
         
+        var appointmentDetails = await _appointmentService.RemoveUserFromAppointmentAsync(id, userId, cancellationToken);
         return Ok(appointmentDetails);
     }
     
-    /// <summary>
-    /// Returns ids of a specific user appointments.
-    /// </summary>
-    /// <param name="id"></param>
-    [HttpGet("user/{id}")]
-    [Produces("application/json")]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(List<int>))]
-    public async Task<IActionResult> GetUsersAppointments([FromRoute] int id, CancellationToken cancellationToken)
-    {
-        return Ok(await _appointmentService.GetAppointmentsIdsByUserIdAsync(id, cancellationToken));
-    }
+    // /// <summary>
+    // /// Returns ids of a specific user appointments.
+    // /// </summary>
+    // /// <param name="id"></param>
+    // [HttpGet("user/{id}")]
+    // [Produces("application/json")]
+    // [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(List<int>))]
+    // public async Task<IActionResult> GetUsersAppointments([FromRoute] int id, CancellationToken cancellationToken)
+    // {
+    //     return Ok(await _appointmentService.GetAppointmentsIdsByUserIdAsync(id, cancellationToken));
+    // }
     
     /// <summary>
     /// Deletes a specific organization.
     /// </summary>
     /// <param name="id"></param>
     [HttpDelete("{id}")]
+    [Authorize(Roles = "Admin,SeniorManager,Manager")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
-    public async Task<IActionResult> DeleteOrganization([FromRoute] int id, CancellationToken cancellationToken)
+    public async Task<IActionResult> DeleteAppointment([FromRoute] int id, CancellationToken cancellationToken)
     {
-        await _appointmentService.DeleteAppointmentByIdAsync(id, cancellationToken);
+        try
+        {
+            bool checkAccess = await _permissionService.CheckAppointmentPermissionAccess(User.Identity.Name, id);
+            if (!checkAccess)
+            {
+                return Forbid();
+            }
+        }
+        catch (NotFoundException)
+        {
+            return NoContent();
+        }
         
         return NoContent();
     }
